:hammer: :tractor: refactored tests
