+++ tests/core/test_context.py
