+++ tests/deprecated/test_handle_requires.py
