+++ tests/test_regression.py
@@ -5,10 +5,9 @@ from textwrap import dedent
 
 import fs.path
 from mock import patch
-from moban import file_system
 from moban.main import main
-from nose.tools import eq_
-from fs.opener.parse import parse_fs_url
+
+from . import utils
 
 
 def custom_dedent(long_texts):
@@ -18,10 +17,7 @@ def custom_dedent(long_texts):
     return refined
 
 
-class TestRegression:
-    def setUp(self):
-        self.current = os.getcwd()
-
+class TestRegression(utils.Docs):
     def test_coping_binary_file(self):
         folder = "regr-01-copy-binary-file"
         args = ["moban"]
@@ -86,35 +82,9 @@ class TestRegression:
     def _raw_moban_with_fs(self, args, folder, expected, output):
         base_dir = fs.path.join("tests", "regression_tests")
         os.chdir(fs.path.join(base_dir, folder))
-        with patch.object(sys, "argv", args):
-            main()
-            _verify_content_with_fs(output, expected)
-        result = parse_fs_url(output)
-        os.unlink(result.resource)  # delete the zip file
+        utils.run_moban_with_fs(args, folder, [(output, expected)])
 
     def _raw_moban_with_fs2(self, args, folder, criterias):
         base_dir = fs.path.join("tests", "regression_tests")
         os.chdir(fs.path.join(base_dir, folder))
-        with patch.object(sys, "argv", args):
-            main()
-
-            for output, expected in criterias:
-                _verify_content_with_fs(output, expected)
-        result = parse_fs_url(output)
-        os.unlink(result.resource)  # delete the zip file
-
-    def tearDown(self):
-        if os.path.exists(".moban.hashes"):
-            os.unlink(".moban.hashes")
-        os.chdir(self.current)
-
-
-def _verify_content(file_name, expected):
-    with open(file_name, "r") as f:
-        content = f.read()
-        eq_(content, expected)
-
-
-def _verify_content_with_fs(file_name, expected):
-    content = file_system.read_unicode(file_name)
-    eq_(content, expected)
+        utils.run_moban_with_fs(args, folder, criterias)
