+++ tests/utils.py
@@ -1,0 +1,46 @@
+import os
+import sys
+
+from mock import patch
+from moban import file_system
+from moban.main import main
+from nose.tools import eq_
+from fs.opener.parse import parse_fs_url
+
+
+def verify_content(file_name, expected):
+    with open(file_name, "r") as f:
+        content = f.read()
+        eq_(content, expected)
+
+
+def verify_content_with_fs(file_name, expected):
+    content = file_system.read_unicode(file_name)
+    eq_(content, expected)
+
+
+def run_moban(args, folder, expected, output):
+    with patch.object(sys, "argv", args):
+        main()
+        verify_content(output, expected)
+    os.unlink(output)
+
+
+def run_moban_with_fs(args, folder, criterias):
+    with patch.object(sys, "argv", args):
+        main()
+
+        for output, expected in criterias:
+            verify_content_with_fs(output, expected)
+    result = parse_fs_url(output)
+    os.unlink(result.resource)  # delete the zip file
+
+
+class Docs:
+    def setUp(self):
+        self.current = os.getcwd()
+
+    def tearDown(self):
+        if os.path.exists(".moban.hashes"):
+            os.unlink(".moban.hashes")
+        os.chdir(self.current)
